+++ assets/javascripts/kitten/components/links/link-list.js
@@ -21,20 +21,32 @@ export class LinkList extends Component {
       active,
     } = element
 
-    const { color } = this.props
+    const {
+      color,
+      lineHeight,
+      itemMargin,
+    } = this.props
 
     const linkListClassName = classNames(
       'k-LinkList__link',
       {
         'is-active': active,
-        'k-LinkList__link--light': color ==  'light',
-        'k-LinkList__link--dark': color ==  'dark',
+        'k-LinkList__link--light': color == 'light',
+        'k-LinkList__link--dark': color == 'dark',
+        'k-LinkList__link--normalLineHeight': lineHeight == 'normal',
+      },
+    )
+
+    const linkListItemClassName = classNames(
+      'k-LinkList__item',
+      {
+        'k-LinkList__item--doubleMargin': itemMargin == 'double',
       },
     )
 
     return (
       <li
-        className="k-LinkList__item"
+        className={ linkListItemClassName }
         key={ key }
       >
         <a
@@ -52,6 +64,8 @@ export class LinkList extends Component {
       className,
       margin,
       items,
+      lineHeight,
+      itemMargin,
       ...others,
     } = this.props
 
@@ -76,6 +90,8 @@ export class LinkList extends Component {
 
 LinkList.defaultProps = {
   color: PropTypes.oneOf(['light', 'dark']),
+  lineHeight: PropTypes.oneOf(['normal']),
+  itemMargin: PropTypes.oneOf(['double']),
 }
 
 LinkList.defaultProps = {
@@ -83,4 +99,6 @@ LinkList.defaultProps = {
   margin: true,
   items: [], // Eg: [{ key: …, item: …, href: … }]
   color: 'dark',
+  lineHeight: null,
+  itemMargin: null,
 }